+++ src/pirate/hwi2c_pio.h
@@ -22,8 +22,8 @@ void pio_i2c_cleanup(void);
 hwi2c_status_t pio_i2c_start_timeout(uint32_t timeout);
 hwi2c_status_t pio_i2c_stop_timeout(uint32_t timeout);
 hwi2c_status_t pio_i2c_restart_timeout(uint32_t timeout);
-hwi2c_status_t pio_i2c_write_timeout(uint32_t out_data, uint32_t timeout);
-hwi2c_status_t pio_i2c_read_timeout(uint32_t* in_data, bool ack, uint32_t timeout);
+hwi2c_status_t pio_i2c_write_timeout(uint8_t out_data, uint32_t timeout);
+hwi2c_status_t pio_i2c_read_timeout(uint8_t* in_data, bool ack, uint32_t timeout);
 hwi2c_status_t pio_i2c_read_array_timeout(uint8_t addr, uint8_t* rxbuf, uint len, uint32_t timeout);
 hwi2c_status_t pio_i2c_write_array_timeout(uint8_t addr, uint8_t* txbuf, uint len, uint32_t timeout);
 hwi2c_status_t pio_i2c_transaction_array_timeout(
